Support CodeUnits as valid input for CSV.File/CSV.Rows (#905)
Fixes #894. I believe this used to work because we made a copy of the
input `CodeUnits` object, which is fine, because we got a
`Vector{UInt8}` out of it, but not ideal since we made a copy. After
some research, I found that when you call `IOBuffer(str)`, it uses this
nifty little trick of calling `unsafe_wrap(Vector{UInt8}, str)` which is
then passed to `IOBuffer` as a way to convert a string to a
`Vector{UInt8}` without copying. We can utilize the same trick to
efficiently treat a string as a `Vector{UInt8}` while the Julia
internals takes care of tracking the true owner of the data as the
original string for us (thus avoiding any GC issues if we were to
naively call `unsafe_wrap(Array, pointer(str))` ourselves).