Allow passing function for types/pool keyword arguments (#906)
* Allow passing function for types/pool keyword arguments

Implements #896. This makes the `types` and `pool` keyword arguments
more consistent with the `select` and `drop` keyword args in that they
accept functions of the form `(col_index, col_name) -> T`, where `types`
should return a type or `nothing`, which will result in the column type
being detected automatically. Similarly for `pool`, the function should
return a `Bool`, `Float64`, or `nothing`.

* update docs

* fix implementation