Try to ensure refpools are promoted between string/inlinestring
Attempted fix at #907. I'm still trying to come up with a way to
reproduce the error, but this may be a missing check where a column type
is ultimately promoted to `String`, but some chunk may have parsed as a
smaller InlineString type and thus needs to be promoted to `String`.