Enhancements to  vector of inputs to CSV.File
This adds a `source` keyword argument when a vector of inputs is
provided that allows specifying a column name that will be added for
each partition of the inputs. It also adds extensive column promotion
functionality/tests for the various column types CSV.File currently
produces.

This still needs some cleanup in tests and additional tests to cover the
`source` keyword arg, but this is pretty close.