Ensure tempfile for compressed input is removed when uncompressed
As astutely discovered in #917, when the input was an IO/Cmd or some
other input, we first buffered input to a tempfile, then if we
discovered the contents of this tempfile were compressed, we would
uncompress the contents to a 2nd tempfile, which was kept track of and
removed at the end of parsing. The issue is the 1st tempfile of
compressed input was forgotten and not property cleaned up (except when
the final process was closed as per `mktemp` defaults).

Unfortunately it's not straightforward to try and "merge" the two
tempfile bufferings since there's no guaranteed interface for peeking a
the first 2 bytes of an IO/Cmd to rely on to detect compression.