Fix macro hygiene issue in CSV.Rows for custom types
Fixes #929. I admit I don't fully understand what exactly is going wrong
here, but it seems that using the `@unrollcolumns` macro inside the `for
i = 1:cols` loop is somehow getting lost when it tries to reference the
`values` array from the unrolled macro block. I know there have been
various issues like this raised at JuliaLang/julia, but it's easy enough
for us to just pass in `values` an explicit argument to the macro, so
that seems an easy enough work-around for now.