Fix corner case when string promotion produces diff type than promoted
Fixes #932. The core issue here is the scenario where the original
multithreaded parse tasks all parse as a non-string (`Int64`) or string
type (like `String3`), but when the non-string tasks go to re-parse
their chunks as string type, it produces a "wider" string type than was
originally promoted. For example, given the types mentioned above, after
intial parsing, the promoted column type would be `String3`, but when a
task goes to reparse the column, it may detect the column type needs to
be wider like `String7`. The problem is that the multithreaded column
post-processing code assumed the originally promoted column type was
"final" and hence bad things happened when a threaded column type was
wider than the promoted type, which shouldn't be possible.

The fix then is to still promote the column to string, but we need to do
that across all threads first, re-promoting the final column type, and
then proceed with the other column post-processing tasks like ref
syncing, etc.