Fix a few type issues in multithreaded parsing (#944)
* Fix a few type issues in multithreaded parsing

Fixes #939. There ended up being a couple issues here. In the originally
posted issue, the "hang" was just due to a very slow code path where our
`nonstandardtype` call was not correctly ignoring `NeedsTypeDetection`,
which is obviously the most common type for columns when no other
columns are provided. The "hang" was due to the 60_000 columns, all
calling `nonstandardtype`, then `tupcat`, which is a pretty expensive
dynamic tuple type creation call. Fixing `nonstandardtype` to correctly
ignore `NeedsTypeDetection` speeds up the case dramatically and avoids
the `tupcat` call. The other issue found once that "hang" was avoided,
was that the new multithreaded detection code was ignoring the `typemap`
keyword arg. This also fixes that case by piping it through to the
`findrowstarts!` family of functions. Note that the new `downcast=true`
keyword arg can be used instead of the original use of `typemap`.

* fix 32-bit