Avoid promoting inline string types larger than String31
Fixes #945. The core issue here is an inconsistency between the initial
type detection code and the later-on type promotion code while parsing.
The type detection code had a limit setup so that inline string column
types weren't allowed larger than `String31`. The promotion code,
however, allowed inline string types to continue to promote up to
`String255`. That means if a column type was at least initiall detected
as some smaller-than-`String31` type, then later on a value larger than
31 bytes was parsed, it would continue to promote up to larger inline
string types. The change proposed in this PR is to limit the promotion
code to be more consistent with the type detection code: if a parsed
value "overflows" the `String31` type, we'll just promote to a regular
`String` instead of promoting to larger inline string types.