Don't warn on final missing value when EOF has been reached
Fixes #948. As noted in the included code comment, if the last expected
value to be parsed is `missing`, the delimited file may not include a
final newline, so the EOF, while abrupt, can be treated as valid. In
this specific case, we can avoid emitting the warning and just ensure
that column is marked as having a `missing` value parsed.