Improve performance of reading files with duplicate column names
I need to load a file with 30k columns, 10k of these have the same
name. Currently, this is practically impossible because makeunique(),
which produces unique column names, has cubic complexity.

This commit changes the algorithm to use a Dict to quickly look up the
existence of columns and to cache the next numeric suffix used to
uniquify column names.

Care has been taken to ensure that columns are named the same way as
before. To that extent, additional tests were added in the previous
commit.