Overhaul how column pooling works while parsing
Fixes #950. Ok, so there's a lot going on here, let me try to summarize
the motivation and what the changes are doing here:
  * In 0.8 releases, `CSV.File` would parse the 1st 100 rows of a file,
  then do a calculation for whether the # of unique values and provided
  `pool` argument should determine if a column should be pooled or not
  * In 0.9, this was changed since just considering the 1st 100 rows
  tends to not be very representative of large files in terms of # of
  unique values. Instead, we parsed the entire file, and afterwards did
  the calculation for whether the column should be pooled.
  * Additionally, in multithreaded parsing, each thread parsed it's own
  chunk of the file, assuming a column would be pooled. After parsing,
  we needed to "synchronize" the ref codes that would be used by a pooled
  column. This ended up being an expensive operation to recode each
  chunk of a column for the whole file and process and reprocess the
  unique values of each chunk. This was notably expensive when the # of
  unique values was large (10s of thousands of values), as reported in
  #950.

That provides the context for why we needed to change, so what is
changing here:
  * We previously tried to assume a column would be pooled before
  parsing and build up the "pool dict" while parsing. This ended up
  complicating a lot of code: additional code to build the dict,
  complications because the element type doesn't match the column type
  (`UInt32` for ref array, `T` for pool dict key type), and
  complications for all the promotion machinery (needing to
  promote/convert pool values instead of ref values). That's in addition
  to the very complex `syncrefs!` routine, that has been the source of
  more a few bugs.
  * Proposed in this PR is to ignore pooling until post-parsing, where
  we'll start checking unique values and do the calculation for whether
  a column should be pooled or not.
  * Also introduced in this PR is allowing the `pool` keyword argument
  to be a number greater than `1.0`, which will be interpreted as an
  absolute upper limit on the # of unique values allowed before a column
  will switch from pooled to non-pooled. This seems to make sense for
  larger files for several reasons: using a % can result in a really
  large # of unique values allowed, which can be a performance
  bottleneck (though still allowed in this PR), and it provides a way
  for us to "short-circuit" the pooling check post-parsing. If the pool
  of unique values gets larger than the pool upper limit threshold, we
  can immediately abort on building an expensive pool with potentially
  lots of unique values. We update the `pool` keyword argument to be a
  default of `500`. Another note is that a file must have number of rows
  > then `pool` upper limit in order to pool, otherwise, the column
  won't be pooled.

One consideration, and potential optimization, is that if
`stringtype=String` or a column otherwise has a type of `String`, we'll
individually allocate each row `String` during parsing, instead of
getting a natural "interning" benefit we got with the pooling strategies
of 0.8 or previous to this PR. During various rounds of benchmarking
every since before 0.8, one thing that has continued to surface is how
efficient allocating individual strings actually is. So while we could
_potentially_ get a little efficiency by interning string columns while
parsing, it's actually a more minimal benefit that people may guess. The
memory saved by interning gets partly used up by having to keep the
intern pool around, and the extra cost of hashing to check to intern can
be comparable to just allocating a smallish string and setting it in our
parsed array. The other consideration is multithreaded parsing: the
interning benefit isn't as strong when each thread has to maintain its
own intern pool (not as many values to intern against). We could perhaps
explore using a lock-free or spin-lock based dict/set for interning
globally for a column, which may end up providing enough performance
benefit. If we do manage to figure out efficient interning, we could
leverage the intern pool post-parsing when considering whether a column
should be pooled or not.